requirements.txt
numpy>=1.24
pandas>=2.0
scipy>=1.10
rich>=13.0
requests>=2.28
xlsxwriter>=3.0
coinbase-advanced-py>=1.0
flask>=3.0
python-dotenv>=1.0
